feat: Improve search input alignment.
Changing the padding around the search field allows us to better support built-in browser functionality, like the "x" clear input button in webkit-based browsers, at the expense of a small aesthetic cost (the focus styling isn't as elegant.)

We might be able to fix this with some Javascript, and apply styling to the container element instead of the actual form input, but for now this seems like a reasonable compromise that prioritises accessibility and usability over sheer aesthetics.

Close #64.